refactor(app): share status drift summary helpers
