+++ openspec/changes/refactor-app-status-drift-summary/README.md
@@ -1,0 +1,3 @@
+# refactor-app-status-drift-summary
+
+Refactor: share status drift summary helpers between CLI and MCP
