+++ src/app/mod.rs
@@ -1,3 +1,4 @@
 pub(crate) mod next_actions;
 pub(crate) mod operator_assets;
 pub(crate) mod preview_diff;
+pub(crate) mod status_drift;
